+++ types/vote_set_test.go
@@ -167,7 +167,7 @@ func Test2_3MajorityRedux(t *testing.T) {
 
 	blockHash := crypto.CRandBytes(32)
 	blockPartsTotal := 123
-	blockPartsHeader := PartSetHeader{blockPartsTotal, crypto.CRandBytes(32)}
+	blockPartsHeader := PartSetHeader{uint32(blockPartsTotal), crypto.CRandBytes(32)}
 
 	voteProto := &Vote{
 		ValidatorAddress: nil, // NOTE: must fill in
@@ -211,7 +211,7 @@ func Test2_3MajorityRedux(t *testing.T) {
 	{
 		addr := privValidators[67].GetPubKey().Address()
 		vote := withValidator(voteProto, addr, 67)
-		blockPartsHeader := PartSetHeader{blockPartsTotal, crypto.CRandBytes(32)}
+		blockPartsHeader := PartSetHeader{uint32(blockPartsTotal), crypto.CRandBytes(32)}
 		_, err := signAddVote(privValidators[67], withBlockPartsHeader(vote, blockPartsHeader), voteSet)
 		if err != nil {
 			t.Error(err)
@@ -226,7 +226,7 @@ func Test2_3MajorityRedux(t *testing.T) {
 	{
 		addr := privValidators[68].GetPubKey().Address()
 		vote := withValidator(voteProto, addr, 68)
-		blockPartsHeader := PartSetHeader{blockPartsTotal + 1, blockPartsHeader.Hash}
+		blockPartsHeader := PartSetHeader{uint32(blockPartsTotal + 1), blockPartsHeader.Hash}
 		_, err := signAddVote(privValidators[68], withBlockPartsHeader(vote, blockPartsHeader), voteSet)
 		if err != nil {
 			t.Error(err)
